+++ cvat-canvas/src/typescript/canvasModel.ts
@@ -60,7 +60,6 @@ export interface DrawData {
     numberOfPoints?: number;
     initialState?: any;
     crosshair?: boolean;
-    // annotation_type?: string;
 }
 
 export interface EditData {