+++ cvat-canvas/src/typescript/canvasView.ts
@@ -40,7 +40,6 @@ import {
     Size,
     Configuration,
 } from './canvasModel';
-import { isNullOrUndefined } from 'util';
 
 export interface CanvasView {
     html(): HTMLDivElement;
@@ -117,14 +116,13 @@ export class CanvasViewImpl implements CanvasView, Listener {
     }
 
     private onEditDone(state: any, points: number[]): void {
-        if (state && points && annotation_type)  {
+        if (state && points)  {
             const event: CustomEvent = new CustomEvent('canvas.edited', {
                 bubbles: false,
                 cancelable: true,
                 detail: {
                     state,
                     points,
-                    // annotation_type,
                 },
             });
 
@@ -445,11 +443,9 @@ export class CanvasViewImpl implements CanvasView, Listener {
                 }
                 if (e.ctrlKey) {
                     const { points } = state;
-                    // const annotation_type = 'Manual';
                     self.onEditDone(
                         state,
                         points.slice(0, pointID * 2).concat(points.slice(pointID * 2 + 2)),
-                        // annotation_type,
                     );
                 } else if (e.shiftKey) {
                     self.canvas.dispatchEvent(new CustomEvent('canvas.editstart', {
@@ -1023,7 +1019,6 @@ export class CanvasViewImpl implements CanvasView, Listener {
             attributes: { ...state.attributes },
             zOrder: state.zOrder,
             pinned: state.pinned,
-            // annotation_type: state.annotation_type,
             updated: state.updated,
             frame: state.frame,
         };
@@ -1380,7 +1375,6 @@ export class CanvasViewImpl implements CanvasView, Listener {
                             id: state.clientID,
                         },
                     }));
-                    // const annotation_type = 'Manual';
                     this.onEditDone(state, points);
                 }
             });
@@ -1429,7 +1423,6 @@ export class CanvasViewImpl implements CanvasView, Listener {
                         + `${shape.attr('x') + shape.attr('width')},`
                         + `${shape.attr('y') + shape.attr('height')}`,
                 ).map((x: number): number => x - offset);
-                // const annotation_type = 'Manual';
 
                 this.drawnStates[state.clientID].points = points;
                 this.canvas.dispatchEvent(new CustomEvent('canvas.resizeshape', {
@@ -1522,14 +1515,14 @@ export class CanvasViewImpl implements CanvasView, Listener {
 
     private addText(state: any): SVG.Text {
         const { undefinedAttrValue } = this.configuration;
-        const { label, clientID, attributes, annotation_type } = state;
+        const { label, clientID, attributes, annotationType } = state;
         const attrNames = label.attributes.reduce((acc: any, val: any): void => {
             acc[val.id] = val.name;
             return acc;
         }, {});
 
         return this.adoptedText.text((block): void => {
-            block.tspan(`${label.name} ${clientID} (${annotation_type})`).style('text-transform', 'uppercase');
+            block.tspan(`${label.name} ${clientID} (${annotationType})`).style('text-transform', 'uppercase');
             for (const attrID of Object.keys(attributes)) {
                 const value = attributes[attrID] === undefinedAttrValue
                     ? '' : attributes[attrID];
@@ -1544,9 +1537,6 @@ export class CanvasViewImpl implements CanvasView, Listener {
 
     private addRect(points: number[], state: any): SVG.Rect {
         const [xtl, ytl, xbr, ybr] = points;
-        // if (typeof (state.annotation_type) === 'undefined') {
-        //     state.annotation_type = 'Manual';
-        // }
         const rect = this.adoptedContent.rect().size(xbr - xtl, ybr - ytl).attr({
             clientID: state.clientID,
             'color-rendering': 'optimizeQuality',
@@ -1556,7 +1546,6 @@ export class CanvasViewImpl implements CanvasView, Listener {
             stroke: state.color,
             'stroke-width': consts.BASE_STROKE_WIDTH / this.geometry.scale,
             'data-z-order': state.zOrder,
-            // annotation_type: state.annotation_type,
         }).move(xtl, ytl)
             .addClass('cvat_canvas_shape');
 
@@ -1572,9 +1561,6 @@ export class CanvasViewImpl implements CanvasView, Listener {
     }
 
     private addPolygon(points: string, state: any): SVG.Polygon {
-        // if (typeof (state.annotation_type) === 'undefined') {
-        //     state.annotation_type = 'Manual';
-        // }
         const polygon = this.adoptedContent.polygon(points).attr({
             clientID: state.clientID,
             'color-rendering': 'optimizeQuality',
@@ -1584,7 +1570,6 @@ export class CanvasViewImpl implements CanvasView, Listener {
             stroke: state.color,
             'stroke-width': consts.BASE_STROKE_WIDTH / this.geometry.scale,
             'data-z-order': state.zOrder,
-            // annotation_type: state.annotation_type,
         }).addClass('cvat_canvas_shape');
 
         if (state.occluded) {
@@ -1599,9 +1584,6 @@ export class CanvasViewImpl implements CanvasView, Listener {
     }
 
     private addPolyline(points: string, state: any): SVG.PolyLine {
-        // if (typeof (state.annotation_type) === 'undefined') {
-        //     state.annotation_type = 'Manual';
-        // }
         const polyline = this.adoptedContent.polyline(points).attr({
             clientID: state.clientID,
             'color-rendering': 'optimizeQuality',
@@ -1611,7 +1593,6 @@ export class CanvasViewImpl implements CanvasView, Listener {
             stroke: state.color,
             'stroke-width': consts.BASE_STROKE_WIDTH / this.geometry.scale,
             'data-z-order': state.zOrder,
-            // annotation_type: state.annotation_type,
         }).addClass('cvat_canvas_shape');
 
         if (state.occluded) {
@@ -1674,14 +1655,12 @@ export class CanvasViewImpl implements CanvasView, Listener {
     }
 
     private addPoints(points: string, state: any): SVG.PolyLine {
-        // state.annotation_type = 'Manual';
         const shape = this.adoptedContent.polyline(points).attr({
             'color-rendering': 'optimizeQuality',
             'pointer-events': 'none',
             'shape-rendering': 'geometricprecision',
             'stroke-width': 0,
             fill: state.color, // to right fill property when call SVG.Shape::clone(
-            // annotation_type: state.annotation_type,
         }).style({
             opacity: 0,
         });