+++ cvat-canvas/src/typescript/drawHandler.ts
@@ -267,11 +267,9 @@ export class DrawHandlerImpl implements DrawHandler {
 
             if (this.canceled) return;
             if ((xbr - xtl) * (ybr - ytl) >= consts.AREA_THRESHOLD) {
-                // this.drawData.annotation_type = 'Manual';
                 this.onDrawDone({
                     shapeType,
                     points: [xtl, ytl, xbr, ybr],
-                    // annotation_type: this.drawData.annotation_type,
                 }, Date.now() - this.startTimestamp);
             }
         }).on('drawupdate', (): void => {
@@ -302,11 +300,9 @@ export class DrawHandlerImpl implements DrawHandler {
                     this.cancel();
 
                     if ((xbr - xtl) * (ybr - ytl) >= consts.AREA_THRESHOLD) {
-                        // this.drawData.annotation_type = 'Manual';
                         this.onDrawDone({
                             shapeType,
                             points: [xtl, ytl, xbr, ybr],
-                            // annotation_type: this.drawData.annotation_type,
                         }, Date.now() - this.startTimestamp);
                     }
                 }
@@ -391,29 +387,23 @@ export class DrawHandlerImpl implements DrawHandler {
             if (shapeType === 'polygon'
                 && ((box.xbr - box.xtl) * (box.ybr - box.ytl) >= consts.AREA_THRESHOLD)
                 && points.length >= 3 * 2) {
-                // this.drawData.annotation_type = 'Manual';
                 this.onDrawDone({
                     shapeType,
                     points,
-                    // annotation_type: this.drawData.annotation_type,
                 }, Date.now() - this.startTimestamp);
             } else if (shapeType === 'polyline'
                 && ((box.xbr - box.xtl) >= consts.SIZE_THRESHOLD
                 || (box.ybr - box.ytl) >= consts.SIZE_THRESHOLD)
                 && points.length >= 2 * 2) {
-                // this.drawData.annotation_type = 'Manual';
                 this.onDrawDone({
                     shapeType,
                     points,
-                    // annotation_type: this.drawData.annotation_type,
                 }, Date.now() - this.startTimestamp);
             } else if (shapeType === 'points'
                 && (e.target as any).getAttribute('points') !== '0,0') {
-                // this.drawData.annotation_type = 'Manual';
                 this.onDrawDone({
                     shapeType,
                     points,
-                    // annotation_type: this.drawData.annotation_type,
                 }, Date.now() - this.startTimestamp);
                 // TODO: think about correct constraign for cuboids
             } else if (shapeType === 'cuboid'
@@ -481,7 +471,6 @@ export class DrawHandlerImpl implements DrawHandler {
 
             if (!e.detail.originalEvent.ctrlKey) {
                 this.release();
-                // this.drawData.annotation_type = 'Manual';
             }
 
             this.onDrawDone({
@@ -492,7 +481,6 @@ export class DrawHandlerImpl implements DrawHandler {
                 attributes: { ...this.drawData.initialState.attributes },
                 label: this.drawData.initialState.label,
                 color: this.drawData.initialState.color,
-                // annotation_type: this.drawData.annotation_type,
             }, Date.now() - this.startTimestamp, e.detail.originalEvent.ctrlKey);
         });
     }
@@ -526,7 +514,6 @@ export class DrawHandlerImpl implements DrawHandler {
             const [xtl, ytl, xbr, ybr] = this.getFinalRectCoordinates(bbox);
             if (!e.detail.originalEvent.ctrlKey) {
                 this.release();
-                // this.drawData.annotation_type = 'Manual';
             }
 
             this.onDrawDone({
@@ -537,7 +524,6 @@ export class DrawHandlerImpl implements DrawHandler {
                 attributes: { ...this.drawData.initialState.attributes },
                 label: this.drawData.initialState.label,
                 color: this.drawData.initialState.color,
-                // annotation_type: this.drawData.annotation_type,
             }, Date.now() - this.startTimestamp, e.detail.originalEvent.ctrlKey);
         });
     }
@@ -650,7 +636,6 @@ export class DrawHandlerImpl implements DrawHandler {
         if (this.drawData.initialState) {
             const { offset } = this.geometry;
             if (this.drawData.shapeType === 'rectangle') {
-                // this.drawData.annotation_type = 'Manual';
                 const [xtl, ytl, xbr, ybr] = this.drawData.initialState.points
                     .map((coord: number): number => coord + offset);
 
@@ -679,7 +664,6 @@ export class DrawHandlerImpl implements DrawHandler {
         } else {
             if (this.drawData.shapeType === 'rectangle') {
                 if (this.drawData.rectDrawingMethod === RectDrawingMethod.EXTREME_POINTS) {
-                    // this.drawData.annotation_type = 'Manual';
                     // draw box by extreme clicking
                     this.drawBoxBy4Points();
                 } else {